+++ packages/vitepress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/vitepress
 
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@0.16.0
+
 ## 0.2.13
 
 ### Patch Changes